+++ tests/Unit/Application/Controller/Admin/AdminDetailsTest.php
@@ -32,19 +32,6 @@ class AdminDetailsTest extends \OxidTestCase
     }
 
     /**
-     * Test get plain editor.
-     */
-    public function testGetPlainEditor()
-    {
-        $oObject = new stdClass();
-        $sEditorHtml = "<textarea id='editor_sField' name='sField' style='width:100px; height:100px;'>sEditObjectValue</textarea>";
-
-        $oAdminDetails = $this->getMock(\OxidEsales\Eshop\Application\Controller\Admin\AdminDetailsController::class, array('_getEditValue'));
-        $oAdminDetails->expects($this->once())->method('_getEditValue')->with($this->equalTo($oObject), $this->equalTo('sField'))->will($this->returnValue('sEditObjectValue'));
-        $this->assertEquals($sEditorHtml, $oAdminDetails->UNITgetPlainEditor(100, 100, $oObject, 'sField'));
-    }
-
-    /**
      * Test get edit value then object is not set.
      */
     public function testGetEditValueObjectNotSet()